# docs/openapi.json
{
  "components": {
    "schemas": {
      "CalibrationRunBody": {
        "additionalProperties": false,
        "description": "A request to score the holdout with the configured jury.",
        "properties": {
          "graded_by": {
            "default": "unknown",
            "minLength": 1,
            "title": "Graded By",
            "type": "string"
          }
        },
        "title": "CalibrationRunBody",
        "type": "object"
      },
...
      "GradeBody": {
        "additionalProperties": false,
        "description": "One grade the author is recording.",
        "properties": {
          "criterion": {
            "minLength": 1,
            "title": "Criterion",
            "type": "string"
          },
          "grade": {
            "maximum": 7.0,
            "minimum": 1.0,
            "title": "Grade",
            "type": "integer"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "sample_id": {
            "minLength": 1,
            "title": "Sample Id",
            "type": "string"
          }
        },
        "required": [
          "sample_id",
          "criterion",
          "grade"
        ],
        "title": "GradeBody",
        "type": "object"
      },
      "GradesBody": {
        "additionalProperties": false,
        "description": "A batch of grades. Partial submission is normal; progress survives interruption.",
        "properties": {
          "graded_by": {
            "default": "unknown",
            "minLength": 1,
            "title": "Graded By",
            "type": "string"
          },
          "grades": {
            "items": {
              "$ref": "#/components/schemas/GradeBody"
            },
            "title": "Grades",
            "type": "array"
          }
        },
        "title": "GradesBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JuryValidationBody": {
        "additionalProperties": false,
        "description": "A jury configuration to dry-run.",
        "properties": {
          "candidate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          }
        },
        "title": "JuryValidationBody",
        "type": "object"
      },
...
      "SampleBody": {
        "additionalProperties": false,
        "description": "One candidate output to be graded: its text, or the run sample to promote.\n\n``content`` is optional because a promoted sample's text is FreeWeight's own record: with\n``source_sample_id`` the service reads the stored response and refuses a ``content`` that\ndiffers from it.",
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "goal_task_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Task Key"
          },
          "origin": {
            "default": "pasted",
            "title": "Origin",
            "type": "string"
          },
          "source_sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Sample Id"
          }
        },
        "title": "SampleBody",
        "type": "object"
      },
      "SamplesBody": {
        "additionalProperties": false,
        "description": "A batch of candidate outputs.",
        "properties": {
          "samples": {
            "items": {
              "$ref": "#/components/schemas/SampleBody"
            },
            "title": "Samples",
            "type": "array"
          }
        },
        "title": "SamplesBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FreeWeight",
    "version": "1.2.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/goals/{slug}/calibration": {
      "get": {
        "description": "Return one goal's calibration samples, the partition, and what remains to be graded.",
        "operationId": "calibration_state_api_v1_goals__slug__calibration_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Calibration State Api V1 Goals  Slug  Calibration Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The calibration set and grading progress",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/v1/goals/{slug}/calibration/grades": {
      "post": {
        "description": "Record the author's grades. Idempotent per ``(sample, criterion)``.",
        "operationId": "submit_grades_api_v1_goals__slug__calibration_grades_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Grades Api V1 Goals  Slug  Calibration Grades Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit grades",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/v1/goals/{slug}/calibration/grading": {
      "get": {
        "description": "What a grader is shown: the samples blinded and shuffled, the criteria and scales, progress.\n\nThe same view FreeWeight's own grading screen renders, and nothing more: a sample's origin,\nits partition, the model that wrote it and every jury grade stay out, because\n:func:`~freeweight.services.calibration.blinded_samples` does not read them out.",
        "operationId": "calibration_grading_api_v1_goals__slug__calibration_grading_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Calibration Grading Api V1 Goals  Slug  Calibration Grading Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The blinded grading view",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/v1/goals/{slug}/calibration/report": {
      "get": {
        "description": "Return the stored report: every coefficient with its ``n_holdout``, and the gate verdict.",
        "operationId": "calibration_report_api_v1_goals__slug__calibration_report_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Calibration Report Api V1 Goals  Slug  Calibration Report Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The stored agreement report",
        "tags": [
          "calibration"
        ]
      }
    },
...
    "/api/v1/goals/{slug}/calibration/run": {
      "post": {
        "description": "Score the held-out samples with the configured jury and store the report.\n\nA rubric that does not clear the gate returns ``200`` with ``calibration_state`` set to\n``\"uncalibrated\"`` and the diagnostics attached: that is a real and useful answer, and\nrefusing the request would withhold exactly what the author needs. Too *few* grades is\n``CALIBRATION_INSUFFICIENT`` — a different state with a different remedy.",
        "operationId": "run_calibration_endpoint_api_v1_goals__slug__calibration_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationRunBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Calibration Endpoint Api V1 Goals  Slug  Calibration Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score the holdout and measure agreement",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/v1/goals/{slug}/calibration/samples": {
      "post": {
        "description": "Add candidate outputs for the author to grade: pasted text, or promoted run samples.\n\nA sample whose content is already present is skipped: two identical samples would be graded\ntwice and counted twice in a figure that assumes independent observations. A promoted sample\nmust be a completed sample of a run of this goal, and its text is read from FreeWeight's own\nrecord.",
        "operationId": "add_calibration_samples_api_v1_goals__slug__calibration_samples_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SamplesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add samples to grade",
        "tags": [
          "calibration"
        ]
      }
    },
...
    "/api/v1/judges": {
      "get": {
        "description": "List every model that may serve as a juror, with the reason for each refusal.",
        "operationId": "list_judges_api_v1_judges_get",
        "parameters": [
          {
            "in": "query",
            "name": "candidate",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Judges Api V1 Judges Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models eligible to judge",
        "tags": [
          "calibration"
        ]
      }
    },
    "/api/v1/judges/validate": {
      "post": {
        "description": "Report the jury that would be assembled, and every refusal with its reason.\n\nNothing is generated: assembling a jury is a decision about eligibility. Self-judging\nconflicts and remote permission are both visible in the response.",
        "operationId": "validate_jury_api_v1_judges_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JuryValidationBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Validate Jury Api V1 Judges Validate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dry-run a jury configuration",
        "tags": [
          "calibration"
        ]
      }
    },
...
  }
}